move routes out of main
